fix: remove duplicate function in environment-resource-service.js
